refactor: Update profilePicture property to require a value
This commit updates the `user.ts` file to change the `profilePicture` property of the `IUser` interface to be required, removing the possibility of it being null. This change ensures that every user has a profile picture associated with their account. It improves the consistency and reliability of the user data.

Closes #456